requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pandas==2.3.1
pyarrow==21.0.0
openpyxl==3.1.5
requests==2.32.4
python-dotenv==1.1.1
pydantic==2.11.7
tqdm==4.67.1
SQLAlchemy==2.0.41